+++ create_sys_dir.sh
@@ -1,0 +1,18 @@
+#!/bin/bash
+
+BASEDIR=$(dirname "$0")
+
+
+create_sys_dir() {
+    dirs=('/tmp/taxhub/' '/tmp/usershub/' "${BASEDIR}/var/log/" "${BASEDIR}/var/log/installdb/")
+    for i in ${dirs[@]}
+    do
+        if [ ! -d $i ]
+        then
+            parentdir="$(dirname "$i")"
+            mkdir -p $i
+        fi
+    done
+    
+    echo "Répertoires système créés"
+}